tests/requirements.py: remove the test_wrapper functions
When a test fails, Pytest prints the failing test function (with an arrow
pointing to the line where the failure occurred). However, with the previous
implementation of `mark_requirement`/`mark_bug`, Pytest considers
`test_wrapper` as the test function, and so it prints the code of
`test_wrapper`, which is useless to the developers.

I rewrote the implementation of `mark_requirement` and `mark_bug` so that an
extra wrapper function is no longer created. I also took the opportunity to
factor out the decorators applied in both functions.