Convert most enum definitions from the functional style to the class style
The class style has advantages in terms of IDE support. In particular, with
Visual Studio Code, when the functional style is used:

* The enumeration type is considered to be a variable rather than a class.
* Autocomplete doesn't work for enum members.
* "Go to definition" doesn't work for enum members (and indeed, with the
  functional style they don't really have proper definitions).

Using the class style solves these problems. I would also argue that the
class style is slightly more readable due to not as much punctuation
(although I can't deny that the repetition of "= auto()" is quite ugly).

The class style removes the need to repeat the enum name twice (in fact,
there were a few cases in the codebase in which the two names didn't match).

I converted most of the enums automatically by using the following script:

```python

import regex
import sys

from pathlib import Path

RE_ENUM = regex.compile(
   r'''
   ^(?<Indent>[ \t]*) (?<EnumName>\w+) \s* = \s* Enum \s* \( \s* (?: '.+?' | ".+?") \s* ,
       \s* \[ \s* (?: '(?<MemberName>.+?)' | "(?<MemberName>.+?)" )
       ( \s* , \s* (?: '(?<MemberName>.+?)' | "(?<MemberName>.+?)" ) )* (?: \s* , )?
       \s* \] \s* \) [ \t]*\n
   ''',
   regex.VERBOSE | regex.MULTILINE)

def main():
    for path in sys.argv[1:]:
        path = Path(path)

        contents = path.read_text()

        def replace(match):
            indent = match.group('Indent')
            enum_name = match.group('EnumName')
            member_names = match.captures('MemberName')
            replacement = f'{indent}class {enum_name}(Enum):\n' \
                + ''.join(f'{indent}    {name} = auto()\n' for name in member_names)
            print(f'{match.group()!r} -> {replacement!r}')
            return replacement

        contents = RE_ENUM.sub(replace, contents)

        path.write_text(contents)

main()
```

The rest of them I converted manually.

The only enum that I kept in the functional style is `MotLabel`, which has
members that are not valid Python identifiers (and probably shouldn't be an
enum at all).