Workflows: install most Python packages at once
In general, pip cannot ensure that after you run multiple `install`
commands, you will still have a consistent environment. Dependencies
installed as part of the first run may be incompatible with packages
installed as part of the second run, in which case pip will install
different versions, which might end up being incompatible with the packages
from the first run.

Installing all packages at once means that pip can figure out which
dependencies can satisfy all packages, and fail if it's impossible.

Unfortunately, setuptools and wheel have to be manually upgraded/installed
for this to work. A recent version of setuptools is required to build
`pycocotools`. wheel is technically not required, but without it,
dependencies that don't have wheels and aren't PEP-517-compliant fall back
to the legacy `setup.py install` path, which might be deprecated soon (see
https://github.com/pypa/pip/issues/8368).

I also replaced `tensorflow` with the `tf` extra on the Datumaro package,
because that way a) it's clearer why we're installing it, and b) we test
that the extra is configured correctly.