Fix an invalid call to pycocotools.mask.iou
The `pyiscrowd` parameter is supposed to have the same length as `gt`, which
in this case is 1, not 2. This mistake used to be harmless, but an assert
for this was added in the upstream repository (see
ppwwyyxx/cocoapi@4f034008), so if a new release of pycocotools comes out,
the current version will start crashing.