imagenet_txt_format: implement automatic category creation (#489)
* imagenet_txt_format: implement automatic category creation

I implemented this as part of prototyping AC integration. AC is able to load
ImageNet without a labels file, and while you can notionally do that with
Datumaro by supplying an empty label list, it doesn't work in practice. As
soon as Datumaro encounters the first annotation, it fails, since the label
ID fails the range check.

Solve this by adding an option to automatically create new categories with
generic names if an out-of-range label ID is encountered.

Also, add a way to specify an empty label list from the command line.

* Add a changelog entry