Refactor Environment._find_plugins (#523)
* Refactor Environment._find_plugins

Previously, it would return a list of tuples of three elements, but two of
those elements would be useless:

* the first element would always be equal to `plugins_dir`;

* the third element would only be used as the `package` argument to
  `importlib.import_module`, which is only needed for relative imports.

In addition, the second element would be a weird cross between a module name
and a file path: it would be dot-separated, yet still have a `.py`
extension.

Replace it with a new implementation that simply returns a list of
dot-separated module names. The new implementation also skips invalid module
names.

* Complain when we see a plugin with a dot in the path

Clarify in the documentation that a plugin must be a module and not just any
file.