Add a new dataset format detection mechanism
Currently, format detection uses the `Importer.find_sources` method.
(Technically, it uses the `Importer.detect` method, but that only has
two implementations, and both of them use `find_sources`.) There are
a couple of problems with that:

1. A lot of `find_sources` implementations aren't very picky and simply
   define a source with the initial directory as the location. As a result,
   autodetection will always return many formats, most of which have nothing
   to do with the dataset.

2. The interface of `find_sources` doesn't allow for easy reporting of
   _why_ the dataset was not detected as a certain format (and AFAIK,
   none of the implementations report it). This information could be useful
   to users.

As a solution, remove the old `detect` method and introduce a new one, which
receives an object that is tailor-made for making checks against a dataset
and reporting any failures. Currently, there is only a generic implementation
that uses `find_sources`, but now we can add custom `detect` implementations
that perform specific checks.

The reporting currently only goes to the debug log. We might want to make it
more prominent in the future.