Modify the detection tests to avoid direct `detect` calls
I'm going to revamp the dataset detection mechanism and make the `detect`
method not as easy to call directly. Therefore, change the detection tests
to instead use `Environment.detect_dataset` and to make sure that the
expected format is one of the detected formats.

There is another benefit to this. The revamped detection mechanism will
allow for more specific checks, and should result in most test datasets
being uniquely detected as belonging to a single format. With the new test
structure, we can easily change the check from
`assertIn(format, detected_formats)` to
`assertEqual([format], detected_formats)` in order to ensure that no other
detectors match the dataset.