Make it easier to access a plugin's name (#525)
CliPlugin._get_name is a cumbersome and seemingly private API for something
that should be easily accessible. For classes that override their plugin name
you can already access it by `cls.NAME`; make that available for all classes
by defining a class attribute in `CliPlugin`.

Use this feature to unhardcode the extractor plugin names in importer
implementations. This way, the link between importers and extractors
is made more explicit.