format_detection: add an ability to place a requirement on file contents
Currently, such requirements are described by a single requirement string
(the one specified in the call to `probe_text_file`). I considered making it
so that you could specify a separate requirement string for each test done
in the prober context (e.g. "must be an XML file"; "must have `annotations`
as the root element"), but that seems cumbersome to use and not terribly
important. If needed, this functionality could be added later (for example,
we could add a method on the context that will tell it to use a more specific
message for the next exception thrown).