Re-add tests for parsing ambiguous revpaths
Previously the subtests in `test_can_parse` checked what happens if a
revpath with no format specified was ambiguous due to multiple detected
formats. However, since the addition of a precise detector for the Datumaro
format, that revpath is not ambiguous anymore. Add a separate test that uses
a dataset that deliberately mixes annotations from different formats.