Add more dataset format detectors (#536)
* format_detection: add an ability to place a requirement on file contents

Currently, such requirements are described by a single requirement string
(the one specified in the call to `probe_text_file`). I considered making it
so that you could specify a separate requirement string for each test done
in the prober context (e.g. "must be an XML file"; "must have `annotations`
as the root element"), but that seems cumbersome to use and not terribly
important. If needed, this functionality could be added later (for example,
we could add a method on the context that will tell it to use a more specific
message for the next exception thrown).

* Add precise detectors for several formats

Unfortunately, JSON files can't really be iteratively parsed (because
object keys can be stored in any order), so when we need to probe the
contents of such files, we have to parse the entire file.

* Give some file contents requirements more specific wording

* format_detection: replace requirement_str with requirement_desc

It's more descriptive.

* Re-add tests for parsing ambiguous revpaths

Previously the subtests in `test_can_parse` checked what happens if a
revpath with no format specified was ambiguous due to multiple detected
formats. However, since the addition of a precise detector for the Datumaro
format, that revpath is not ambiguous anymore. Add a separate test that uses
a dataset that deliberately mixes annotations from different formats.