site/build_docs.py: behave less self-destructively
Currently, it overwrites files in the actual repo it is run from, which is
inconvenient and dangerous (since any local changes are lost). Make it stage
files in a temporary directory instead.

This requires the `enableGitInfo` setting to be set to `false`, since we no
longer run Hugo in a Git repository. This makes the "Last modified"
information disappear from the generated pages. However, due to the way we
currently generate documentation for old releases, that information is
frequently incorrect anyway (for example:
<https://openvinotoolkit.github.io/datumaro/v0.1.11/docs/contributing/>).