Make various improvements to site/build_docs.py (#589)
* site/build_docs.py: make executable

* site/build_docs.py: miscellaneous improvements

* site/build_docs.py: don't create output directories manually

It seems that Hugo is perfectly capable of creating them by itself.

A positive side effect of this change is that you can now run the script
repeatedly (previously it would complain that the output directory already
exists).

* site/build_docs.py: switch to using pathlib

Also, use the actual repository root as the repository root, not the current
directory. Change the variable names accordingly.

* site/build_docs.py: fail if hugo fails

* site/build_docs.py: behave less self-destructively

Currently, it overwrites files in the actual repo it is run from, which is
inconvenient and dangerous (since any local changes are lost). Make it stage
files in a temporary directory instead.

This requires the `enableGitInfo` setting to be set to `false`, since we no
longer run Hugo in a Git repository. This makes the "Last modified"
information disappear from the generated pages. However, due to the way we
currently generate documentation for old releases, that information is
frequently incorrect anyway (for example:
<https://openvinotoolkit.github.io/datumaro/v0.1.11/docs/contributing/>).