ByteImage: try to guess the image extension if one is not given
This lets us avoid needless re-encoding when a `ByteImage` is saved as
the format it already is in.

Unfortunately, neither OpenCV nor PIL seem to offer functionality to detect
the format without loading the image, so we have to implement it ourselves.
I only added support for a handful of most common formats, which should cover
a majority of cases.