Implement public dataset downloading via TFDS (#582)
* Add an extra for TensorFlow Datasets

This will be used by a future `download` command.

Note that TFDS does not explicitly depend on TensorFlow, and neither does
this extra. This is because there are multiple TF distributions
(`tensorflow`, `tensorflow-gpu`, `tf-nightly`), and not having an explicit
dependency allows the user to decide which distribution to install.

* Add a command to download publicly available datasets

In terms of syntax & semantics, this command is comparable to the `convert`
command, except that instead of a source directory and source format, it
accepts a "dataset ID".

Currently, the downloading is done through an external library, TensorFlow
Datasets. However, I expect that this might not always be the case: we might
implement native downloading later, or add other download backends.
Therefore, the dataset ID must begin with a namespace (currently, the only
such namespace is `tfds:`) that signifies the download backend.

We could probably make use of the Datumaro plugin manager functionality to
implement download backends, but I don't want to bother with it until there
are at least two of them.

The way TFDS represents dataset items is much more flexible than the way
Datumaro does it; an item can have features with arbitrary names and types.
To map those features onto Datumaro's fixed annotation types, this code uses
adapters (which are basically just sequences of predefined callbacks that
each convert a certain feature or a metadata element into the Datumaro
representation).

* Add tests for the dataset download functionality

* Add documentation for the `download` command

* download: Print a more specific error message when TFDS is not installed

* Add a changelog entry

* download: Mention the proxy environment variables in the documentation

* download: make the format parameter optional

Use the format of the original dataset by default.

Rename `_TfdsAdapter.metadata_transformers` to `category_transformers` to
avoid confusion with `metadata`.