FormatRequirementsUnmet: replace generate_message with __str__
`__str__` is the typical way to report exception messages, and I don't see
a compelling reason not to use it here.

Also, change the `human_message` parameter of `RejectionCallback` to an
`str`, since there's no real need to report the message as a sequence of
lines rather than as a multiline string. Also, with the sequence
representation, it was possible to produce a line with a `\n` in the middle,
which is undesirable. A single multiline string is not susceptible to this.