Add a dedicated command to detect a dataset's format (#576)
* Move most of `Environment.detect_dataset` to `format_detection`

There is actually very little in this code that deals with environments,
and this move makes it possible to reuse the code for other purposes.

* Introduce a format detection progress reporter class

This will allow verbose reporting of format detection outcomes.

* Add a new detect-format command

* Add a docstring to detect_dataset_format

* Add tests for the detect-format command

* Add documentation for the detect-format command

* detect_dataset_format: simplify reporter interface

I don't think it's actually necessary to report different rejection reasons
via different methods, given how similar they are represented in the JSON
report. If we ever need to pass custom data for different reasons, we can
modify the callback / add a new one.

* Add a changelog entry

* FormatRequirementsUnmet: replace generate_message with __str__

`__str__` is the typical way to report exception messages, and I don't see
a compelling reason not to use it here.

Also, change the `human_message` parameter of `RejectionCallback` to an
`str`, since there's no real need to report the message as a sequence of
lines rather than as a multiline string. Also, with the sequence
representation, it was possible to produce a line with a `\n` in the middle,
which is undesirable. A single multiline string is not susceptible to this.

* Rewrite the docstring of `detect_dataset_format` in a more formal style

Also, rename the `detectors` parameter to `formats`, since that's a better
representation of its semantics.

* Change an exception variable name to a more conventional one

I think the old name was derived from a class name that is no longer used.
It's irrelevant now.

* Add tests for the detect_dataset_format function