voc_format: don't create XML files with no objects on export
The original VOC datasets (at least the 2007 and 2012 versions that I have
checked) do not contain XML files with no defined objects. Where an image
is not annotated with objects, the corresponding XML file is simply omitted.

This change makes the Datumaro VOC converter consistent with that convention
by avoiding dumping the XML files when there are no objects associated with
a dataset item.

Beside consistency, the other benefit of this is that written datasets can
be disambiguated from the LabelMe format by the format detection machinery.
The VOC and LabelMe formats seem to only differ by how annotations are
represented inside the `object` element, so if there are no objects, the
resulting XML file is ambiguous. By avoiding writing such XML files, we
avoid the ambiguity.