labelme_format: try harder to disambiguate from VOC
A dataset in the VOC format may contain XML files with no objects, which
will be indistinguishable from LabelMe files. Instead of matching
immediately in this case, keep checking files until either one of them can
be definitively determined to be in the LabelMe format or not, or all files
are exhausted.