format_detection: add a dedicated function to report unsupported detection (#665)
Currently, instances where detection for a format is unsupported are
reported via `context.fail`. However, `context.fail` is intended primarily
for situations where the dataset fails some requirement, and using it for
this leads to unnatural-looking messages:

    The following requirement was not met:
      this format cannot be autodetected

Fix this by adding a dedicated rejection reason for this case and a new
method, `context.raise_unsupported` to abort detection for this reason.