labelme_format: try harder to disambiguate from VOC (#664)
* labelme_format: try harder to disambiguate from VOC

A dataset in the VOC format may contain XML files with no objects, which
will be indistinguishable from LabelMe files. Instead of matching
immediately in this case, keep checking files until either one of them can
be definitively determined to be in the LabelMe format or not, or all files
are exhausted.

* test_format_detection: remove redundant result asserts

There already is a dedicated test that checks the default result value,
so it's not necessary to check it elsewhere.