Save box attributes when downloading the TFDS version of VOC
The main nuisance with this is that Datumaro represents the pose as a
string, while TFDS represents it as an index, and to convert the latter into
the former, we need access to the list of poses.

To implement this, allow adapters access to a persistent state object that
can be used to store and retrieve arbitrary data during conversion. We can
then save the list of names during category creation and retrieve it when we
convert box annotations.