Remove usages of distutils from the library
`distutils` has been deprecated in Python 3.10. Fortunately, the only name
we're using from there is `strtobool`, and conveniently, `attrs` has
recently added a function with the same semantics and the same sets of true
and false values. So just use that instead.

Note that `strtobool` is still used in `setup.py`, and we can't use `attrs`
there. I'm hoping that setuptools will eventually provide a replacement.