extractor_tfds: title-case VOC pose names
TFDS returns the pose names as lower case, but they are upper case in the
original dataset. Convert them to title case so that the output of the
extractor better resembles the original dataset.

Remove `_SaveFeatureClassList`, since it's not useful anymore (and it's
unclear whether it'll ever be useful again).