Save box attributes when downloading the TFDS version of VOC (#668)
* Remove usages of distutils from the library

`distutils` has been deprecated in Python 3.10. Fortunately, the only name
we're using from there is `strtobool`, and conveniently, `attrs` has
recently added a function with the same semantics and the same sets of true
and false values. So just use that instead.

Note that `strtobool` is still used in `setup.py`, and we can't use `attrs`
there. I'm hoping that setuptools will eventually provide a replacement.

* Save box attributes when downloading the TFDS version of VOC

The main nuisance with this is that Datumaro represents the pose as a
string, while TFDS represents it as an index, and to convert the latter into
the former, we need access to the list of poses.

To implement this, allow adapters access to a persistent state object that
can be used to store and retrieve arbitrary data during conversion. We can
then save the list of names during category creation and retrieve it when we
convert box annotations.

* extractor_tfds: title-case VOC pose names

TFDS returns the pose names as lower case, but they are upper case in the
original dataset. Convert them to title case so that the output of the
extractor better resembles the original dataset.

Remove `_SaveFeatureClassList`, since it's not useful anymore (and it's
unclear whether it'll ever be useful again).