Fix the GitHub pages workflow (#699)
This workflow uses Ubuntu 20.04, which has pip 20.0.2 installed by default.
However, the latest version of the `openvino` wheel uses the
`manylinux_2_27_x86_64` platform tag, which was not supported until pip
20.3. pip 20.0.2 also doesn't have the new resolver enabled by default, so
it's not smart enough to try an older version of `openvino-dev`. Therefore,
this pip cannot install `openvino` and therefore `openvino-dev`.

This can be solved by upgrading pip, and there _is_ a pip upgrade command in
the workflow. However, it has no effect, because the upgrade installs the
new version as `/usr/local/bin/pip`, while bash remembers that pip was at
`/usr/bin/pip` and keeps using that version.

Fix it by separating the pip upgrade into a separate step. It looks nicer
this way in the build log, anyway.