COCO: save skeleton points in the correct order if they are labelled (#14)
The COCO format does not support storing points out of order; they must be
stored in the same order as they are listed in the skeleton definition. If
the points in the dataset are not labelled, we can only assume they are
already in the right order; but if they are, we can use the labels to
rearrange them into the correct order before saving.

If we don't do this, and the dataset has the points in an arbitrary order
(which can happen when it is converted from CVAT), then the points will be
written in that arbitrary order, making the original mapping of
point-to-label no longer recoverable.