Add ImageColorScale context manager (#1194)
- To utilize Datumaro as the data pipeline frontend for model training,
it is crucial to ensure that the input image's color format consistently
adheres to the 3-channel requirement. This is a common practice for many
model training frameworks. This pull request introduces a context
manager designed to enforce the acquisition of raw images in Datumaro
(e.g., data.media_as(Image).data) in the 3-channel.

Signed-off-by: Kim, Vinnam <vinnam.kim@intel.com>